+++ test/test_table.cpp
@@ -7082,4 +7082,59 @@ TEST(Table_getVersionCounterAfterRowAccessor)
     _CHECK_VER_BUMP();
 }
 
+
+// This test a bug where get_size_from_type_and_ref() returned off-by-one on nullable integer columns.
+// It seems to be only invoked from Table::get_size_from_ref() which is fast static method that lets
+// you find the size of a Table without having to create an instance of it. This seems to be only done
+// on subtables, so the bug has not been triggered in public.
+TEST_TYPES(Table_ColumnSizeFromRef, std::true_type, std::false_type)
+{
+    constexpr bool nullable = TEST_TYPE::value;
+    Group g;
+    TableRef t = g.add_table("table");
+    t->add_column(type_Int, "int", nullable);
+    t->add_column(type_Bool, "bool", nullable);
+    t->add_column(type_String, "string", nullable);
+    t->add_column(type_Binary, "binary", nullable);
+    t->add_column(type_Double, "double");
+    t->add_column(type_Float, "float");
+    t->add_column(type_Mixed, "mixed");
+    t->add_column(type_Timestamp, "timestamp");
+    t->add_column_link(type_Link, "link", *t);
+    t->add_column_link(type_LinkList, "LinkList", *t);
+
+    auto check_column_sizes = [this, &t](size_t num_rows) {
+        t->clear();
+        t->add_empty_row(num_rows);
+        CHECK_EQUAL(t->size(), num_rows);
+        using tf = _impl::TableFriend;
+        Spec& t_spec = tf::get_spec(*t);
+        size_t actual_num_cols = t_spec.get_column_count();
+        for (size_t col_ndx = 0; col_ndx < actual_num_cols; ++col_ndx) {
+            ColumnType col_type = t_spec.get_column_type(col_ndx);
+            ColumnBase& base = tf::get_column(*t, col_ndx);
+            ref_type col_ref = base.get_ref();
+            bool nullable = (t_spec.get_column_attr(col_ndx) & col_attr_Nullable) == col_attr_Nullable;
+            size_t col_size = ColumnBase::get_size_from_type_and_ref(col_type, col_ref, base.get_alloc(), nullable);
+            CHECK_EQUAL(col_size, num_rows);
+        }
+    };
+
+    // Test leafs
+    check_column_sizes(REALM_MAX_BPNODE_SIZE - 1);
+
+    // Test empty
+    check_column_sizes(0);
+
+    // Test internal nodes
+    check_column_sizes(REALM_MAX_BPNODE_SIZE + 1);
+
+    // Test on boundary for good measure
+    check_column_sizes(REALM_MAX_BPNODE_SIZE);
+
+    // Try with more levels in the tree
+    check_column_sizes(10 * REALM_MAX_BPNODE_SIZE);
+}
+
+
 #endif // TEST_TABLE